agent1/requirements.txt
pandas==2.1.4
geopandas==0.14.2
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
geoalchemy2==0.14.3
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
shapely==2.0.2
pyproj==3.6.1
unidecode==1.3.7
anthropic==0.52.0
loguru==0.7.2
tenacity==8.2.3
scikit-learn==1.3.2

# Phase 3 — Self-Healing Data Pipeline (LangChain ReAct agents)
langchain>=0.3.21
langchain-anthropic>=0.3.14
langchain-community>=0.3.20
duckduckgo-search>=6.3.7

